requirements/dev.txt
# Everything the developer needs in addition to the production requirements
-r prod.txt

# Testing
pytest
mock
coverage
factory-boy==2.4.0

# Style checkers/linters
flake8
pep8

# Management script
Flask-Script

honcho
